reorganise app-context to separate different segments
