+++ client/app-context/SheetyContext/actions.ts
@@ -1,47 +1,32 @@
-import { TSheetyFacilities, Items, TSheetyMethods, TSheetyCategories } from "api/sheety/types";
+import { Items, TSheetyMethods, TSheetyCategories, TSheetyFacilities } from "api/sheety/types";
 import {
-	Actions,
-	AddressOption,
-	TInitializeCategoriesAction,
 	TInitializeItemsListAction,
+	SheetyActionsEnums,
 	TInitializeMethodsAction,
-	TItemSelection,
-	TSetAddressAction,
-	TSetUserSelectionAction,
+	TInitializeCategoriesAction,
 	TInitializeFacilitiesAction,
-	TEmptyItem,
 } from "./types";
 
 const initializeItemsList = (items: Items[]): TInitializeItemsListAction => ({
-	type: Actions.SET_ITEMS_LIST,
+	type: SheetyActionsEnums.SET_ITEMS_LIST,
 	payload: items,
 });
 
 const initializeMethodsList = (methods: TSheetyMethods[]): TInitializeMethodsAction => ({
-	type: Actions.SET_METHODS_LIST,
+	type: SheetyActionsEnums.SET_METHODS_LIST,
 	payload: methods,
 });
 
 const initializeCategoriessList = (
 	categories: TSheetyCategories[],
 ): TInitializeCategoriesAction => ({
-	type: Actions.SET_CATEGORIES_LIST,
+	type: SheetyActionsEnums.SET_CATEGORIES_LIST,
 	payload: categories,
 });
 
 const initializeFaciltiesList = (facilities: TSheetyFacilities[]): TInitializeFacilitiesAction => ({
-	type: Actions.SET_FACILITIES_LIST,
+	type: SheetyActionsEnums.SET_FACILITIES_LIST,
 	payload: facilities,
-});
-
-const setUserSelection = (selection: (TItemSelection | TEmptyItem)[]): TSetUserSelectionAction => ({
-	type: Actions.SET_USER_SELECTION,
-	payload: selection,
-});
-
-const setAddress = (address: AddressOption): TSetAddressAction => ({
-	type: Actions.SET_ADDRESS,
-	payload: address,
 });
 
 export const SheetyActions = {
@@ -49,9 +34,4 @@ export const SheetyActions = {
 	initializeCategoriessList,
 	initializeMethodsList,
 	initializeFaciltiesList,
-};
-
-export const SelectionActions = {
-	setUserSelection,
-	setAddress,
 };
