+++ client/app-context/SheetyContext/index.ts
@@ -1,0 +1,11 @@
+import { SheetyContextState } from "./types";
+
+export const initialSheetyState: SheetyContextState = {
+	recyclableItems: {
+		isLoaded: false,
+		data: [],
+	},
+	methods: [],
+	categories: [],
+	facilities: [],
+};
