+++ client/app-context/UserSelectionContext/index.ts
@@ -1,0 +1,20 @@
+import { UserSelectionContextState } from "./types";
+
+export const initialUserSelectionState: UserSelectionContextState = {
+	userSelection: {
+		address: {
+			value: "",
+			label: "",
+			coordinates: {
+				lat: "",
+				long: "",
+			},
+		},
+		items: [
+			{
+				name: "",
+				method: undefined,
+			},
+		],
+	},
+};
