+++ client/app-context/reducer.ts
@@ -1,10 +1,12 @@
 import { Reducer } from "react";
-import { Actions, AppContextActions, AppContextState } from "./types";
+import { AppContextActions, AppContextState } from "./types";
 import { transformCategory, transformFacility } from "./utils";
+import { SheetyActionsEnums } from "./SheetyContext/types";
+import { UserSelectionActionsEnums } from "./UserSelectionContext/types";
 
 export const AppContextReducer: Reducer<AppContextState, AppContextActions> = (state, action) => {
 	switch (action.type) {
-		case Actions.SET_ITEMS_LIST:
+		case SheetyActionsEnums.SET_ITEMS_LIST:
 			return {
 				...state,
 				recyclableItems: {
@@ -12,22 +14,22 @@ export const AppContextReducer: Reducer<AppContextState, AppContextActions> = (s
 					data: action.payload,
 				},
 			};
-		case Actions.SET_METHODS_LIST:
+		case SheetyActionsEnums.SET_METHODS_LIST:
 			return {
 				...state,
 				methods: action.payload.map((method) => method.name),
 			};
-		case Actions.SET_CATEGORIES_LIST:
+		case SheetyActionsEnums.SET_CATEGORIES_LIST:
 			return {
 				...state,
 				categories: action.payload.map((category) => transformCategory(category)),
 			};
-		case Actions.SET_FACILITIES_LIST:
+		case SheetyActionsEnums.SET_FACILITIES_LIST:
 			return {
 				...state,
 				facilities: action.payload.map((facility) => transformFacility(facility)),
 			};
-		case Actions.SET_USER_SELECTION:
+		case UserSelectionActionsEnums.SET_USER_SELECTION:
 			return {
 				...state,
 				userSelection: {
@@ -35,7 +37,7 @@ export const AppContextReducer: Reducer<AppContextState, AppContextActions> = (s
 					items: action.payload,
 				},
 			};
-		case Actions.SET_ADDRESS:
+		case UserSelectionActionsEnums.SET_ADDRESS:
 			return {
 				...state,
 				userSelection: {
